Refactor to component-based architecture
